requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas
python-dotenv==1.0.1
ollama==0.3.3
google-generativeai==0.8.2
matplotlib==3.9.2
streamlit==1.39.0
plotly==5.24.1
openai==1.51.0